+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## v1.1.0
-This release introduces the Pampas upgrade and contains 6 new features.  
+This release introduces the Pampas upgrade and contains 8 new features.  
 
 Features:
 * [#316](https://github.com/bnb-chain/greenfield-cosmos-sdk/pull/316) feat: add validation with context information
@@ -10,6 +10,8 @@ Features:
 * [#341](https://github.com/bnb-chain/greenfield-cosmos-sdk/pull/341) feat: add support for some json rpc queries
 * [#353](https://github.com/bnb-chain/greenfield-cosmos-sdk/pull/353) feat: distinguish inturn relayer
 * [#320](https://github.com/bnb-chain/greenfield-cosmos-sdk/pull/320) feat: add MsgRejectMigrateBucket
+* [#357](https://github.com/bnb-chain/greenfield-cosmos-sdk/pull/357) feat: support Secp256k1 format private keys import
+* [#358](https://github.com/bnb-chain/greenfield-cosmos-sdk/pull/358) feat: enable Pampas hardfork to testnet
 
 ## v1.0.1
 This release includes 1 bug fix.